requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
PyJWT==2.7.0
python-decouple==3.8
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
twilio==8.5.0
typing-extensions==4.7.1
urllib3==2.0.3
yarl==1.9.2
